+++ databricks/koalas/data_type_ops.py
@@ -19,10 +19,13 @@ Classes for binary operations between Koalas objects, classified by object data 
 """
 from abc import ABCMeta, abstractmethod
 import datetime
-import numpy as np
 import warnings
 
+import numpy as np
+import pandas as pd
 from pandas.api.types import CategoricalDtype
+
+from pyspark import sql as spark
 from pyspark.sql import Column, functions as F
 from pyspark.sql.types import (
     BooleanType,
@@ -40,7 +43,7 @@ from pyspark.sql.types import (
 
 from databricks.koalas.base import column_op, IndexOpsMixin, numpy_column_op
 from databricks.koalas.spark import functions as SF
-from databricks.koalas.typedef import Dtype, as_spark_type
+from databricks.koalas.typedef import Dtype, as_spark_type, extension_dtypes
 
 
 class DataTypeOps(object, metaclass=ABCMeta):
@@ -138,6 +141,62 @@ class DataTypeOps(object, metaclass=ABCMeta):
     @abstractmethod
     def __rmod__(self, left, right=None):
         raise NotImplementedError()
+
+    def __and__(self, left, right):
+        if isinstance(left.dtype, extension_dtypes) or (
+            isinstance(right, IndexOpsMixin) and isinstance(right.dtype, extension_dtypes)
+        ):
+
+            def and_func(left, right):
+                if not isinstance(right, spark.Column):
+                    if pd.isna(right):
+                        right = F.lit(None)
+                    else:
+                        right = F.lit(right)
+                return left & right
+
+        else:
+
+            def and_func(left, right):
+                if not isinstance(right, spark.Column):
+                    if pd.isna(right):
+                        right = F.lit(None)
+                    else:
+                        right = F.lit(right)
+                scol = left & right
+                return F.when(scol.isNull(), False).otherwise(scol)
+
+        return column_op(and_func)(left, right)
+
+    def __rand__(self, left, right=None):
+        return self.__and__(left, right)
+
+    def __or__(self, left, right):
+        if isinstance(left.dtype, extension_dtypes) or (
+            isinstance(right, IndexOpsMixin) and isinstance(right.dtype, extension_dtypes)
+        ):
+
+            def or_func(left, right):
+                if not isinstance(right, spark.Column):
+                    if pd.isna(right):
+                        right = F.lit(None)
+                    else:
+                        right = F.lit(right)
+                return left | right
+
+        else:
+
+            def or_func(left, right):
+                if not isinstance(right, spark.Column) and pd.isna(right):
+                    return F.lit(False)
+                else:
+                    scol = left | F.lit(right)
+                    return F.when(left.isNull() | scol.isNull(), False).otherwise(scol)
+
+        return column_op(or_func)(left, right)
+
+    def __ror__(self, left, right=None):
+        return self.__or__(left, right)
 
 
 class NumericOps(DataTypeOps):
